Update content security policy
Firefox changed the requirements for the content security policy. The result was
that the extension did not load anymore. Only an empty popup was displayed and the
option page was also empty.

New content security policy allows eval as this is required by Vue.js as used in
the extension.