Correct migration to new settings
When the plugin is run for the first time, the settings
are not initialized and the tabs were stored in the old
storage format. This leads to a migration on the first
loading.

In addition, on the first load after migration, the
storage key is calculated incorrectly, because tab
group information is missing. Therefor no tabs are
loaded. When loading again, tabs were loaded correctly.

Now settings are initialized correctly and the tabs
are loaded on the first attempt.